refactor: set proper namespace
