+++ perception/detected_object_validation/src/obstacle_pointcloud/debugger.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DETECTED_OBJECT_VALIDATION__OBSTACLE_POINTCLOUD_BASED_VALIDATOR__DEBUGGER_HPP_
-#define DETECTED_OBJECT_VALIDATION__OBSTACLE_POINTCLOUD_BASED_VALIDATOR__DEBUGGER_HPP_
+#ifndef OBSTACLE_POINTCLOUD__DEBUGGER_HPP_
+#define OBSTACLE_POINTCLOUD__DEBUGGER_HPP_
 
 #include <rclcpp/rclcpp.hpp>
 
@@ -24,7 +24,9 @@
 #include <pcl/point_types.h>
 #include <pcl_conversions/pcl_conversions.h>
 
-namespace obstacle_pointcloud_based_validator
+namespace autoware::detected_object_validation
+{
+namespace obstacle_pointcloud
 {
 class Debugger
 {
@@ -106,6 +108,8 @@ private:
     return pointcloud_xyz;
   }
 };
-}  // namespace obstacle_pointcloud_based_validator
 
-#endif  // DETECTED_OBJECT_VALIDATION__OBSTACLE_POINTCLOUD_BASED_VALIDATOR__DEBUGGER_HPP_
+}  // namespace obstacle_pointcloud
+}  // namespace autoware::detected_object_validation
+
+#endif  // OBSTACLE_POINTCLOUD__DEBUGGER_HPP_
