+++ perception/detected_object_validation/test/object_position_filter/test_object_position_filter.cpp
@@ -21,10 +21,10 @@
 
 #include <vector>
 
+using autoware::detected_object_validation::position_filter::ObjectPositionFilterNode;
 using autoware_perception_msgs::msg::DetectedObject;
 using autoware_perception_msgs::msg::DetectedObjects;
 using autoware_perception_msgs::msg::ObjectClassification;
-using object_position_filter::ObjectPositionFilterNode;
 
 std::shared_ptr<autoware::test_utils::AutowareTestManager> generateTestManager()
 {
